Create NS testing output

Create testing output similar to when running a 'BEAST2' run with nested sampling

Usage

create_test_ns_output()

See also

Author

Richèl J.C. Bilderbeek

Examples

beastier::remove_beaustier_folders()
beastier::check_empty_beaustier_folders()

create_test_ns_output()
#>   [1] "[/usr/lib/jvm/java-8-oracle/bin/java, -Djava.library.path=/usr/local/lib:/usr/java/packages/lib/amd64:/usr/lib64:/lib64:/lib:/usr/lib, -cp, ::/home/richel/.beast/2.5/MODEL_SELECTION/lib/MODEL_SELECTION.addon.jar:/home/richel/.beast/2.5/BEASTLabs/lib/BEASTlabs.addon.jar:/home/richel/.beast/2.5/NS/lib/NS.addon.jar:/home/richel/.beast/2.5/BEAST/lib/beast.src.jar:/home/richel/.beast/2.5/BEAST/lib/beast.jar, beast.app.beastapp.BeastMain, nested_sampling.xml]"
#>   [2] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>   [3] "                        BEAST v2.5.0, 2002-2018"                                                                                                                                                                                                                                                                                                                                                                                                                          
#>   [4] "             Bayesian Evolutionary Analysis Sampling Trees"                                                                                                                                                                                                                                                                                                                                                                                                               
#>   [5] "                       Designed and developed by"                                                                                                                                                                                                                                                                                                                                                                                                                         
#>   [6] " Remco Bouckaert, Alexei J. Drummond, Andrew Rambaut & Marc A. Suchard"                                                                                                                                                                                                                                                                                                                                                                                                   
#>   [7] "                                    "                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>   [8] "                     Department of Computer Science"                                                                                                                                                                                                                                                                                                                                                                                                                      
#>   [9] "                         University of Auckland"                                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [10] "                        remco@cs.auckland.ac.nz"                                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [11] "                        alexei@cs.auckland.ac.nz"                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [12] "                                    "                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [13] "                   Institute of Evolutionary Biology"                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [14] "                        University of Edinburgh"                                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [15] "                           a.rambaut@ed.ac.uk"                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [16] "                                    "                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [17] "                    David Geffen School of Medicine"                                                                                                                                                                                                                                                                                                                                                                                                                      
#>  [18] "                 University of California, Los Angeles"                                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [19] "                           msuchard@ucla.edu"                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  [20] "                                    "                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [21] "                      Downloads, Help & Resources:"                                                                                                                                                                                                                                                                                                                                                                                                                       
#>  [22] "                           http://beast2.org/"                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [23] "                                    "                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [24] "  Source code distributed under the GNU Lesser General Public License:"                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [25] "                   http://github.com/CompEvol/beast2"                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [26] "                                    "                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [27] "                           BEAST developers:"                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  [28] "   Alex Alekseyenko, Trevor Bedford, Erik Bloomquist, Joseph Heled, "                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [29] " Sebastian Hoehna, Denise Kuehnert, Philippe Lemey, Wai Lok Sibon Li, "                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [30] "Gerton Lunter, Sidney Markowitz, Vladimir Minin, Michael Defoin Platel, "                                                                                                                                                                                                                                                                                                                                                                                                 
#>  [31] "          Oliver Pybus, Tim Vaughan, Chieh-Hsi Wu, Walter Xie"                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [32] "                                    "                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [33] "                               Thanks to:"                                                                                                                                                                                                                                                                                                                                                                                                                                
#>  [34] "          Roald Forsberg, Beth Shapiro and Korbinian Strimmer"                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [35] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [36] "Random number seed: 1539258042094"                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [37] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [38] "File: nested_sampling.xml seed: 1539258042094 threads: 1"                                                                                                                                                                                                                                                                                                                                                                                                                 
#>  [39] "61430_aco: 78 4"                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [40] "626029_aco: 78 4"                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [41] "630116_aco: 78 4"                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [42] "630210_aco: 78 4"                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [43] "B25702_aco: 78 4"                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [44] "Alignment(anthus_aco_sub)"                                                                                                                                                                                                                                                                                                                                                                                                                                                
#>  [45] "  5 taxa"                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#>  [46] "  78 sites"                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#>  [47] "  9 patterns"                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#>  [48] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [49] "Failed to load BEAGLE library: no hmsbeagle-jni in java.library.path"                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [50] "TreeLikelihood(treeLikelihood.anthus_aco_sub0) uses BeerLikelihoodCore4"                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [51] "  Alignment(anthus_aco_sub): [taxa, patterns, sites] = [5, 9, 78]"                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [52] "==============================================================================="                                                                                                                                                                                                                                                                                                                                                                                          
#>  [53] "Citations for this model:"                                                                                                                                                                                                                                                                                                                                                                                                                                                
#>  [54] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [55] "Patricio Maturana, Brendon J. Brewer, Steffen Klaere, Remco Bouckaert. Model selection and parameter inference in phylogenetics using Nested Sampling. Systematic Biology, syy050, 2018"                                                                                                                                                                                                                                                                                  
#>  [56] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [57] "==============================================================================="                                                                                                                                                                                                                                                                                                                                                                                          
#>  [58] "Counting 19 parameters"                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [59] "replacing logger tracelog with NSLogger"                                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [60] "Writing file anthus_aco_sub.trees"                                                                                                                                                                                                                                                                                                                                                                                                                                        
#>  [61] "         Sample      posterior ESS(posterior)     likelihood          prior"                                                                                                                                                                                                                                                                                                                                                                                              
#>  [62] "Writing file anthus_aco_sub.log"                                                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [63] "Start likelihood 0: -225.2488738545589 "                                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [64] "1 particles initialised"                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [65] "              0      -140.6759              N      -157.4874        16.8115 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [66] "ML: -158.32601908408833 Information: 0.8385606384288167"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [67] "              1      -124.4630         2.0         -143.0242        18.5611 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [68] "ML: -144.86277038390745 Information: 1.8385386453449868"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [69] "              2      -124.6624         3.0         -141.6496        16.9872 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [70] "ML: -143.96492404000415 Information: 1.7551898946725117"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [71] "              3      -124.9536         4.0         -139.0497        14.0961 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [72] "ML: -142.59510010333074 Information: 2.7422179576086307"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [73] "             24      -125.0399        14.8         -137.0128        11.9729 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [74] "ML: -141.98047696172614 Information: 2.8349216678190885"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [75] "             25      -124.7614        15.4         -137.0127        12.2512 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [76] "ML: -141.98047696086337 Information: 2.8349216687952605"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [77] "             26      -124.7070        15.9         -137.0124        12.3054 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [78] "ML: -141.98047696054587 Information: 2.8349216691561026"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [79] "             27      -124.6071        16.2         -137.0106        12.4035 --"                                                                                                                                                                                                                                                                                                                                                                                           
#>  [80] "ML: -141.98047696042886 Information: 2.8349216692884056"                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [81] "28<=10000&& (28<2.0*2.8349216692884056*1||Math.abs(-141.98047696042886 - -141.98047696054587)/Math.abs(-141.98047696042886) > 1.0E-12"                                                                                                                                                                                                                                                                                                                                    
#>  [82] "Finished in 28 steps!"                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>  [83] "Marginal likelihood: -141.31236790862764 (bootstrap SD=1.4240215445106443)"                                                                                                                                                                                                                                                                                                                                                                                               
#>  [84] "Marginal likelihood: -140.98848039195855 (subsample SD=1.8638865000383444)"                                                                                                                                                                                                                                                                                                                                                                                               
#>  [85] "Marginal likelihood: -141.1644574217485(1.1601428428211327)"                                                                                                                                                                                                                                                                                                                                                                                                              
#>  [86] "Information: 2.8349216692884056"                                                                                                                                                                                                                                                                                                                                                                                                                                          
#>  [87] "SD: 1.6837225630395305"                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [88] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [89] "Operator                                                    Tuning    #accept    #reject      Pr(m)  Pr(acc|m)"                                                                                                                                                                                                                                                                                                                                                           
#>  [90] "ScaleOperator(YuleBirthRateScaler.t:anthus_aco_sub)        0.75000       4797        998    0.04000    0.82778 Try setting scaleFactor to about 0.562"                                                                                                                                                                                                                                                                                                                    
#>  [91] "ScaleOperator(YuleModelTreeScaler.t:anthus_aco_sub)        0.50000       1116       4748    0.04000    0.19031 "                                                                                                                                                                                                                                                                                                                                                          
#>  [92] "ScaleOperator(YuleModelTreeRootScaler.t:anthus_aco_sub)    0.50000        887       4915    0.04000    0.15288 "                                                                                                                                                                                                                                                                                                                                                          
#>  [93] "Uniform(YuleModelUniformOperator.t:anthus_aco_sub)               -      11869      46191    0.40000    0.20443 "                                                                                                                                                                                                                                                                                                                                                          
#>  [94] "SubtreeSlide(YuleModelSubtreeSlide.t:anthus_aco_sub)       1.00000        128      28742    0.20000    0.00443 Try decreasing size to about 0.5"                                                                                                                                                                                                                                                                                                                          
#>  [95] "Exchange(YuleModelNarrow.t:anthus_aco_sub)                       -       3007      26046    0.20000    0.10350 "                                                                                                                                                                                                                                                                                                                                                          
#>  [96] "Exchange(YuleModelWide.t:anthus_aco_sub)                         -        180       5641    0.04000    0.03092 "                                                                                                                                                                                                                                                                                                                                                          
#>  [97] "WilsonBalding(YuleModelWilsonBalding.t:anthus_aco_sub)           -         92       5643    0.04000    0.01604 "                                                                                                                                                                                                                                                                                                                                                          
#>  [98] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [99] "     Tuning: The value of the operator's tuning parameter, or '-' if the operator can't be optimized."                                                                                                                                                                                                                                                                                                                                                                    
#> [100] "    #accept: The total number of times a proposal by this operator has been accepted."                                                                                                                                                                                                                                                                                                                                                                                    
#> [101] "    #reject: The total number of times a proposal by this operator has been rejected."                                                                                                                                                                                                                                                                                                                                                                                    
#> [102] "      Pr(m): The probability this operator is chosen in a step of the MCMC (i.e. the normalized weight)."                                                                                                                                                                                                                                                                                                                                                                 
#> [103] "  Pr(acc|m): The acceptance probability (#accept as a fraction of the total proposals for this operator)."                                                                                                                                                                                                                                                                                                                                                                
#> [104] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [105] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [106] "Total calculation time: 0.823 seconds"                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> [107] "End likelihood: 12.388056869575024"                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [108] "Producing posterior samples"                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [109] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [110] "Marginal likelihood: -141.3000915386624 sqrt(H/N)=(NaN)=?=SD=(1.4593051179523482) Information: -0.19242514244311712"                                                                                                                                                                                                                                                                                                                                                      
#> [111] "Max ESS: 5.659472090348357"                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [112] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [113] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [114] "Processing 28 trees from file."                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [115] "Log file written to anthus_aco_sub.posterior.trees"                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [116] "Done!"                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> [117] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [118] "Marginal likelihood: -141.31271098696405 sqrt(H/N)=(NaN)=?=SD=(1.462246661656647) Information: -0.1353639346412825"                                                                                                                                                                                                                                                                                                                                                       
#> [119] "Max ESS: 5.4945059635784155"                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [120] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [121] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [122] "Log file written to anthus_aco_sub.posterior.log"                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [123] "Done!"                                                                                                                                                                                                                                                                                                                                                                                                                                                                    

beastier::remove_beaustier_folders()
beastier::check_empty_beaustier_folders()